Añadido el StateHolder para gestionar la lógica del snackbar de ShopDetail
