--- app/src/main/java/dev/rulsoft/oprbattles/club/domain/ClubSource.kt
@@ -1,7 +1,0 @@
-package dev.rulsoft.oprbattles.club.domain
-
-interface ClubSource {
-    suspend fun fetchClubs(): List<Club>
-
-    suspend fun fetchClubById(uid: String): Club?
-}
