--- app/src/main/java/dev/rulsoft/oprbattles/club/presentation/ClubState.kt
@@ -1,8 +1,0 @@
-package dev.rulsoft.oprbattles.club.presentation
-
-import dev.rulsoft.oprbattles.club.presentation.models.ClubUi
-
-data class ClubState(
-    val loading: Boolean = false,
-    val club: ClubUi? = null
-)
