+++ app/src/main/java/dev/rulsoft/oprbattles/core/navigation/graphs/ShopGraph.kt
@@ -6,8 +6,8 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
 import dev.rulsoft.oprbattles.core.navigation.ShopNav
 import dev.rulsoft.oprbattles.core.navigation.ShopsNav
-import dev.rulsoft.oprbattles.shop.presentation.ShopDetailScreen
-import dev.rulsoft.oprbattles.shop.presentation.ShopListScreen
+import dev.rulsoft.oprbattles.presentation.shop.ShopDetailScreen
+import dev.rulsoft.oprbattles.presentation.shop.ShopListScreen
 import kotlinx.serialization.Serializable
 
 @Serializable
