+++ app/src/main/java/dev/rulsoft/oprbattles/data/shop/networking/dto/ShopDto.kt
@@ -1,0 +1,16 @@
+package dev.rulsoft.oprbattles.data.shop.networking.dto
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ShopDto(
+    val uid: String = "",
+    val displayName: String = "",
+    val email: String = "",
+    val photoUrl: String = "",
+    val phone: String = "",
+    val address: String = "",
+    val openingHour: String = "",
+    val manager: String = "",
+    val hasEventSpace: Boolean = false
+)
