+++ app/src/main/java/dev/rulsoft/oprbattles/domain/club/ClubSource.kt
@@ -1,0 +1,7 @@
+package dev.rulsoft.oprbattles.domain.club
+
+interface ClubSource {
+    suspend fun fetchClubs(): List<Club>
+
+    suspend fun fetchClubById(uid: String): Club?
+}
