+++ app/src/main/java/dev/rulsoft/oprbattles/domain/shop/ShopData.kt
@@ -1,0 +1,7 @@
+package dev.rulsoft.oprbattles.domain.shop
+
+interface ShopData {
+    suspend fun fetchShops(): List<Shop>
+
+    suspend fun fetchShopById(uid: String): Shop?
+}
