+++ app/src/main/java/dev/rulsoft/oprbattles/presentation/club/ClubState.kt
@@ -1,0 +1,9 @@
+package dev.rulsoft.oprbattles.presentation.club
+
+import dev.rulsoft.oprbattles.presentation.club.models.ClubUi
+
+data class ClubState(
+    val loading: Boolean = false,
+    val club: ClubUi? = null
+)
+
