+++ app/src/main/java/dev/rulsoft/oprbattles/presentation/shop/ShopListState.kt
@@ -1,0 +1,8 @@
+package dev.rulsoft.oprbattles.presentation.shop
+
+import dev.rulsoft.oprbattles.presentation.shop.models.ShopUi
+
+data class ShopListState (
+    val loading: Boolean = false,
+    val shops: List<ShopUi> = emptyList()
+)
