+++ app/src/main/java/dev/rulsoft/oprbattles/presentation/shop/ShopState.kt
@@ -1,0 +1,9 @@
+package dev.rulsoft.oprbattles.presentation.shop
+
+import dev.rulsoft.oprbattles.presentation.shop.models.ShopUi
+
+data class ShopState(
+    val loading: Boolean = false,
+    val shop: ShopUi? = null,
+    val message: String? = null
+)
