--- app/src/main/java/dev/rulsoft/oprbattles/shop/domain/ShopData.kt
@@ -1,7 +1,0 @@
-package dev.rulsoft.oprbattles.shop.domain
-
-interface ShopData {
-    suspend fun fetchShops(): List<Shop>
-
-    suspend fun fetchShopById(uid: String): Shop?
-}
