--- app/src/main/java/dev/rulsoft/oprbattles/shop/presentation/ShopListState.kt
@@ -1,8 +1,0 @@
-package dev.rulsoft.oprbattles.shop.presentation
-
-import dev.rulsoft.oprbattles.shop.presentation.models.ShopUi
-
-data class ShopListState (
-    val loading: Boolean = false,
-    val shops: List<ShopUi> = emptyList()
-)
