--- app/src/main/java/dev/rulsoft/oprbattles/shop/presentation/ShopState.kt
@@ -1,9 +1,0 @@
-package dev.rulsoft.oprbattles.shop.presentation
-
-import dev.rulsoft.oprbattles.shop.presentation.models.ShopUi
-
-data class ShopState(
-    val loading: Boolean = false,
-    val shop: ShopUi? = null,
-    val message: String? = null
-)
