+++ app/src/main/java/dev/rulsoft/oprbattles/ui/composable/ClubList.kt
@@ -12,8 +12,8 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
-import dev.rulsoft.oprbattles.club.presentation.ClubListViewModel
-import dev.rulsoft.oprbattles.club.presentation.models.ClubUi
+import dev.rulsoft.oprbattles.presentation.club.ClubListViewModel
+import dev.rulsoft.oprbattles.presentation.club.models.ClubUi
 
 @Composable
 fun ClubList(
